task 4 remake

+++ Edge.h
@@ -15,7 +15,7 @@ class Edge {
 public:
     //int s;
     int to = -1;
-    float dist = 1e9;
+    double dist = 1e9;
     //int prev_pos = 0;
     friend istream &operator>>(istream &in, Edge &edge);
     friend ostream &operator<<(ostream &out, const Edge &edge);
